requirements.txt
discord
requests
datetime
psutil
asyncio
pytz
python-dotenv
colorama
typing-extensions